+++ src/main.exe/ReqItemMakibishi.c
@@ -81,7 +81,7 @@
  *    for this one store, same as the other twins.
  */
 extern void ProcItemMakibishi(tag_TItem *item);
-extern s16 SoundEx(VECTOR *loc, int id);
+extern s16 SoundEx(VECTOR *loc, short id);
 /* This TU defines the counter (gp-relative): listed in Build.hs
  * maspsxGpExterns for this file, unlike ActionHalt/FRAMES (absolute here). */
 extern s32 COUNTER_FOR_ITEM_ARRAY_;
